+++ bee-message/src/payload/mod.rs
@@ -28,7 +28,7 @@ use core::{convert::Infallible, fmt};
 
 /// Maximum length (in bytes) of a message payload, defined in the specification:
 /// <https://github.com/iotaledger/IOTA-2.0-Research-Specifications/blob/main/2.3%20Standard%20Payloads%20Layout.md>.
-pub const PAYLOAD_LENGTH_MAX: usize = 65157;
+pub const PAYLOAD_LENGTH_MAX: u32 = 65157;
 
 /// Error encountered packing a payload.
 #[derive(Debug)]
